Add files via upload
Install the application give and it the required permissions. Point to the camera towards a plane surface and click on the screen
nd your as soon as a surface is detected, you will see an virtual object placed there.
NOTE-This is an apk file so it only works on Android devices. Unity also gives you the option to develop for IOS devices.